requirements.txt
flask
pyopenssl
google-cloud-storage
google-cloud-aiplatform
google-cloud-texttospeech